Added support for listening on TCP sockets instead of UNIX sockets.
In the configuration file, add the following:

	clientSocket = "tcp://127.0.0.1:9998"
	eventSocket = "tcp://127.0.0.1:9999"

If not starting with tcp://, it will be treated as a filename, and
UNIX socket will be used.
Defaults are /tmp/TelldusClient and /tmp/TelldusEvents

Only working on UNIX right now, no socket impl in windows, and CF
settings won't work in OSX (I think?).