Added TCP support for libtelldus-core (telldus client).
In order to setup the library with special sockets endpoints, use the
tdInitWithSocketSpec(char*, char*) API call.
Calling the old tdInit() sets up with default sockets paths.

This introduces some internal changes in the client library, the static
methods on the Client class are now gone. In order to keep the address
set in the new init-call, we store the address in the Client instance
which was always created anyway  (getInstance()). Since the old
behaviour was to always create an instance of this object anyway, this
shouldn't introduce any problems.

The external behaviour has not changed at all, except for the new
tdInitWithSocketSpec method.
Currently only on UNIX though, the method is hidden in Windows since
there is no backend code for this yet.