Race-fix for OSX ControllerListener.
If the destructor was called very close (in time) after the constructor
was called, we'd get a race where gRunLoop was not inited properly, and
when trying to release it we crashed with EXC_BAD_ACCESS.

An example when this would happen is if the user SIGINTs while starting
up. The ControllerManager can be quite slow (seconds) to init sometimes,
and if we'd SIGINT while waiting for it, the ControllerListener ctor
would be called after, the mainloop would be skipped, and then the dtor
would be called -> triggering crash.